sqlccl: split restore out of backup.go
I've been wanting to do this but didn't want to mix it with other changes, so here it is as a standalone change.
